+++ src/main/kotlin/com/github/jcornaz/kable/impl/ThreeMapTable.kt
@@ -32,7 +32,7 @@ import com.github.jcornaz.kable.util.toTableEntry
  * * map of row-value maps by columns
  * * map of column-value maps by rows
  */
-class ThreeMapTable<R, C, V>(entries: Iterable<Table.Entry<R, C, V>>) : AbstractTable<R, C, V>(), MutableTable<R, C, V> {
+class ThreeMapTable<R, C, V>(entries: Iterable<Table.Entry<R, C, V>> = emptyList()) : AbstractTable<R, C, V>(), MutableTable<R, C, V> {
 
     private val map = mutableMapOf<Pair<R, C>, V>()
     private val rowsMap = mutableMapOf<R, MutableMap<C, V>>()